requirements.txt
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
imageio==2.15.0
importlib-resources==5.4.0
kiwisolver==1.3.1
matplotlib==3.3.4
networkx==2.5.1
ninja==1.10.2.3
numpy==1.19.5
opencv-python==4.5.5.64
Pillow==8.4.0
pyparsing==3.0.7
python-dateutil==2.8.2
PyWavelets==1.1.1
PyYAML==6.0
scikit-image==0.17.2
scipy==1.5.4
six==1.16.0
tifffile==2020.9.3
torch==1.10.2+cu113
torchaudio==0.10.2+cu113
torchvision==0.11.3+cu113
tqdm==4.64.0
typing_extensions==4.1.1
zipp==3.6.0
